Add chick weights dataset and add documentation to datasets methods
