+++ creme/data/chick-weight.csv
@@ -1,0 +1,579 @@
+weight,time,chick,diet
+42,0,1,1
+40,0,2,1
+43,0,3,1
+42,0,4,1
+41,0,5,1
+41,0,6,1
+41,0,7,1
+42,0,8,1
+42,0,9,1
+41,0,10,1
+43,0,11,1
+41,0,12,1
+41,0,13,1
+41,0,14,1
+41,0,15,1
+41,0,16,1
+42,0,17,1
+39,0,18,1
+43,0,19,1
+41,0,20,1
+40,0,21,2
+41,0,22,2
+43,0,23,2
+42,0,24,2
+40,0,25,2
+42,0,26,2
+39,0,27,2
+39,0,28,2
+39,0,29,2
+42,0,30,2
+42,0,31,3
+41,0,32,3
+39,0,33,3
+41,0,34,3
+41,0,35,3
+39,0,36,3
+41,0,37,3
+41,0,38,3
+42,0,39,3
+41,0,40,3
+42,0,41,4
+42,0,42,4
+42,0,43,4
+42,0,44,4
+41,0,45,4
+40,0,46,4
+41,0,47,4
+39,0,48,4
+40,0,49,4
+41,0,50,4
+51,2,1,1
+49,2,2,1
+39,2,3,1
+49,2,4,1
+42,2,5,1
+49,2,6,1
+49,2,7,1
+50,2,8,1
+51,2,9,1
+44,2,10,1
+51,2,11,1
+49,2,12,1
+48,2,13,1
+49,2,14,1
+49,2,15,1
+45,2,16,1
+51,2,17,1
+35,2,18,1
+48,2,19,1
+47,2,20,1
+50,2,21,2
+55,2,22,2
+52,2,23,2
+52,2,24,2
+49,2,25,2
+48,2,26,2
+46,2,27,2
+46,2,28,2
+48,2,29,2
+48,2,30,2
+53,2,31,3
+49,2,32,3
+50,2,33,3
+49,2,34,3
+53,2,35,3
+48,2,36,3
+48,2,37,3
+49,2,38,3
+50,2,39,3
+55,2,40,3
+51,2,41,4
+49,2,42,4
+55,2,43,4
+51,2,44,4
+50,2,45,4
+52,2,46,4
+53,2,47,4
+50,2,48,4
+53,2,49,4
+54,2,50,4
+59,4,1,1
+58,4,2,1
+55,4,3,1
+56,4,4,1
+48,4,5,1
+59,4,6,1
+57,4,7,1
+61,4,8,1
+59,4,9,1
+52,4,10,1
+63,4,11,1
+56,4,12,1
+53,4,13,1
+62,4,14,1
+56,4,15,1
+49,4,16,1
+61,4,17,1
+55,4,19,1
+54,4,20,1
+62,4,21,2
+64,4,22,2
+61,4,23,2
+58,4,24,2
+62,4,25,2
+57,4,26,2
+58,4,27,2
+58,4,28,2
+59,4,29,2
+59,4,30,2
+62,4,31,3
+65,4,32,3
+63,4,33,3
+63,4,34,3
+64,4,35,3
+61,4,36,3
+56,4,37,3
+61,4,38,3
+61,4,39,3
+66,4,40,3
+66,4,41,4
+63,4,42,4
+69,4,43,4
+65,4,44,4
+61,4,45,4
+62,4,46,4
+66,4,47,4
+62,4,48,4
+64,4,49,4
+67,4,50,4
+64,6,1,1
+72,6,2,1
+67,6,3,1
+67,6,4,1
+60,6,5,1
+74,6,6,1
+71,6,7,1
+71,6,8,1
+68,6,9,1
+63,6,10,1
+84,6,11,1
+62,6,12,1
+60,6,13,1
+79,6,14,1
+64,6,15,1
+51,6,16,1
+72,6,17,1
+62,6,19,1
+58,6,20,1
+86,6,21,2
+77,6,22,2
+73,6,23,2
+74,6,24,2
+78,6,25,2
+74,6,26,2
+73,6,27,2
+73,6,28,2
+74,6,29,2
+72,6,30,2
+73,6,31,3
+82,6,32,3
+77,6,33,3
+85,6,34,3
+87,6,35,3
+76,6,36,3
+68,6,37,3
+74,6,38,3
+78,6,39,3
+79,6,40,3
+85,6,41,4
+84,6,42,4
+96,6,43,4
+86,6,44,4
+78,6,45,4
+82,6,46,4
+79,6,47,4
+80,6,48,4
+85,6,49,4
+84,6,50,4
+76,8,1,1
+84,8,2,1
+84,8,3,1
+74,8,4,1
+79,8,5,1
+97,8,6,1
+89,8,7,1
+84,8,8,1
+85,8,9,1
+74,8,10,1
+112,8,11,1
+72,8,12,1
+65,8,13,1
+101,8,14,1
+68,8,15,1
+57,8,16,1
+83,8,17,1
+65,8,19,1
+65,8,20,1
+125,8,21,2
+90,8,22,2
+90,8,23,2
+66,8,24,2
+102,8,25,2
+93,8,26,2
+87,8,27,2
+92,8,28,2
+87,8,29,2
+85,8,30,2
+85,8,31,3
+107,8,32,3
+96,8,33,3
+107,8,34,3
+123,8,35,3
+98,8,36,3
+80,8,37,3
+98,8,38,3
+89,8,39,3
+101,8,40,3
+103,8,41,4
+103,8,42,4
+131,8,43,4
+103,8,44,4
+98,8,45,4
+101,8,46,4
+100,8,47,4
+104,8,48,4
+108,8,49,4
+105,8,50,4
+93,10,1,1
+103,10,2,1
+99,10,3,1
+87,10,4,1
+106,10,5,1
+124,10,6,1
+112,10,7,1
+93,10,8,1
+96,10,9,1
+81,10,10,1
+139,10,11,1
+88,10,12,1
+67,10,13,1
+128,10,14,1
+68,10,15,1
+51,10,16,1
+89,10,17,1
+71,10,19,1
+73,10,20,1
+163,10,21,2
+95,10,22,2
+103,10,23,2
+68,10,24,2
+124,10,25,2
+114,10,26,2
+100,10,27,2
+114,10,28,2
+106,10,29,2
+98,10,30,2
+102,10,31,3
+129,10,32,3
+111,10,33,3
+134,10,34,3
+158,10,35,3
+116,10,36,3
+83,10,37,3
+109,10,38,3
+109,10,39,3
+120,10,40,3
+124,10,41,4
+126,10,42,4
+157,10,43,4
+118,10,44,4
+117,10,45,4
+120,10,46,4
+123,10,47,4
+125,10,48,4
+128,10,49,4
+122,10,50,4
+106,12,1,1
+122,12,2,1
+115,12,3,1
+102,12,4,1
+141,12,5,1
+141,12,6,1
+146,12,7,1
+110,12,8,1
+90,12,9,1
+89,12,10,1
+168,12,11,1
+119,12,12,1
+71,12,13,1
+164,12,14,1
+67,12,15,1
+54,12,16,1
+98,12,17,1
+82,12,19,1
+77,12,20,1
+217,12,21,2
+108,12,22,2
+127,12,23,2
+70,12,24,2
+146,12,25,2
+136,12,26,2
+115,12,27,2
+145,12,28,2
+134,12,29,2
+115,12,30,2
+123,12,31,3
+159,12,32,3
+137,12,33,3
+164,12,34,3
+201,12,35,3
+145,12,36,3
+103,12,37,3
+128,12,38,3
+130,12,39,3
+154,12,40,3
+155,12,41,4
+160,12,42,4
+184,12,43,4
+127,12,44,4
+135,12,45,4
+144,12,46,4
+148,12,47,4
+154,12,48,4
+152,12,49,4
+155,12,50,4
+125,14,1,1
+138,14,2,1
+138,14,3,1
+108,14,4,1
+164,14,5,1
+148,14,6,1
+174,14,7,1
+116,14,8,1
+92,14,9,1
+96,14,10,1
+177,14,11,1
+135,14,12,1
+70,14,13,1
+192,14,14,1
+68,14,15,1
+103,14,17,1
+88,14,19,1
+89,14,20,1
+240,14,21,2
+111,14,22,2
+135,14,23,2
+71,14,24,2
+164,14,25,2
+147,14,26,2
+123,14,27,2
+156,14,28,2
+150,14,29,2
+122,14,30,2
+138,14,31,3
+179,14,32,3
+144,14,33,3
+186,14,34,3
+238,14,35,3
+166,14,36,3
+112,14,37,3
+154,14,38,3
+146,14,39,3
+182,14,40,3
+153,14,41,4
+174,14,42,4
+188,14,43,4
+138,14,44,4
+141,14,45,4
+156,14,46,4
+157,14,47,4
+170,14,48,4
+166,14,49,4
+175,14,50,4
+149,16,1,1
+162,16,2,1
+163,16,3,1
+136,16,4,1
+197,16,5,1
+155,16,6,1
+218,16,7,1
+126,16,8,1
+93,16,9,1
+101,16,10,1
+182,16,11,1
+162,16,12,1
+71,16,13,1
+227,16,14,1
+113,16,17,1
+106,16,19,1
+98,16,20,1
+275,16,21,2
+131,16,22,2
+145,16,23,2
+72,16,24,2
+197,16,25,2
+169,16,26,2
+144,16,27,2
+184,16,28,2
+187,16,29,2
+143,16,30,2
+170,16,31,3
+221,16,32,3
+151,16,33,3
+235,16,34,3
+287,16,35,3
+198,16,36,3
+135,16,37,3
+192,16,38,3
+170,16,39,3
+215,16,40,3
+175,16,41,4
+204,16,42,4
+197,16,43,4
+145,16,44,4
+147,16,45,4
+173,16,46,4
+168,16,47,4
+222,16,48,4
+184,16,49,4
+205,16,50,4
+171,18,1,1
+187,18,2,1
+187,18,3,1
+154,18,4,1
+199,18,5,1
+160,18,6,1
+250,18,7,1
+134,18,8,1
+100,18,9,1
+112,18,10,1
+184,18,11,1
+185,18,12,1
+81,18,13,1
+248,18,14,1
+123,18,17,1
+120,18,19,1
+107,18,20,1
+307,18,21,2
+148,18,22,2
+163,18,23,2
+72,18,24,2
+231,18,25,2
+205,18,26,2
+163,18,27,2
+207,18,28,2
+230,18,29,2
+151,18,30,2
+204,18,31,3
+263,18,32,3
+146,18,33,3
+294,18,34,3
+332,18,35,3
+227,18,36,3
+157,18,37,3
+232,18,38,3
+214,18,39,3
+262,18,40,3
+184,18,41,4
+234,18,42,4
+198,18,43,4
+146,18,44,4
+174,18,45,4
+210,18,46,4
+185,18,47,4
+261,18,48,4
+203,18,49,4
+234,18,50,4
+199,20,1,1
+209,20,2,1
+198,20,3,1
+160,20,4,1
+220,20,5,1
+160,20,6,1
+288,20,7,1
+125,20,8,1
+100,20,9,1
+120,20,10,1
+181,20,11,1
+195,20,12,1
+91,20,13,1
+259,20,14,1
+133,20,17,1
+144,20,19,1
+115,20,20,1
+318,20,21,2
+164,20,22,2
+170,20,23,2
+76,20,24,2
+259,20,25,2
+236,20,26,2
+185,20,27,2
+212,20,28,2
+279,20,29,2
+157,20,30,2
+235,20,31,3
+291,20,32,3
+156,20,33,3
+327,20,34,3
+361,20,35,3
+225,20,36,3
+169,20,37,3
+280,20,38,3
+250,20,39,3
+295,20,40,3
+199,20,41,4
+269,20,42,4
+199,20,43,4
+197,20,45,4
+231,20,46,4
+210,20,47,4
+303,20,48,4
+233,20,49,4
+264,20,50,4
+205,21,1,1
+215,21,2,1
+202,21,3,1
+157,21,4,1
+223,21,5,1
+157,21,6,1
+305,21,7,1
+98,21,9,1
+124,21,10,1
+175,21,11,1
+205,21,12,1
+96,21,13,1
+266,21,14,1
+142,21,17,1
+157,21,19,1
+117,21,20,1
+331,21,21,2
+167,21,22,2
+175,21,23,2
+74,21,24,2
+265,21,25,2
+251,21,26,2
+192,21,27,2
+233,21,28,2
+309,21,29,2
+150,21,30,2
+256,21,31,3
+305,21,32,3
+147,21,33,3
+341,21,34,3
+373,21,35,3
+220,21,36,3
+178,21,37,3
+290,21,38,3
+272,21,39,3
+321,21,40,3
+204,21,41,4
+281,21,42,4
+200,21,43,4
+196,21,45,4
+238,21,46,4
+205,21,47,4
+322,21,48,4
+237,21,49,4
+264,21,50,4
